Обработка исключений для CategoryTransactionsController

+++ src/main/java/ru/ServerRestApp/controllers/CategoryTransactionsController.java
@@ -5,10 +5,14 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.BindingResult;
-import org.springframework.validation.FieldError;
 import org.springframework.web.bind.annotation.*;
+import ru.ServerRestApp.models.Category;
 import ru.ServerRestApp.models.CategoryTransaction;
+import ru.ServerRestApp.models.Person;
+import ru.ServerRestApp.models.Team;
+import ru.ServerRestApp.services.CategoriesService;
 import ru.ServerRestApp.services.CategoryTransactionsService;
+import ru.ServerRestApp.services.PeopleService;
 import ru.ServerRestApp.util.ErrorResponse;
 import ru.ServerRestApp.util.DataException;
 import ru.ServerRestApp.util.NotFoundException;
@@ -23,46 +27,103 @@ import static ru.ServerRestApp.util.ErrorsUtil.returnDataErrorsToClient;
 public class CategoryTransactionsController {
 
     private final CategoryTransactionsService categoryTransactionsService;
+    private final PeopleService peopleService;
+    private final CategoriesService categoriesService;
     @Autowired
-    public CategoryTransactionsController(CategoryTransactionsService categoryTransactionsService) {
+    public CategoryTransactionsController(CategoryTransactionsService categoryTransactionsService, PeopleService peopleService, CategoriesService categoriesService) {
         this.categoryTransactionsService = categoryTransactionsService;
+        this.peopleService = peopleService;
+        this.categoriesService = categoriesService;
     }
 
 
     @GetMapping()
-    public List<CategoryTransaction> getAllCategoryTransactions() {
-        return categoryTransactionsService.findAll();
+    public ResponseEntity<List<CategoryTransaction>> getAllCategoryTransactions() {
+        List<CategoryTransaction> categoryTransactions = categoryTransactionsService.findAll();
+        return new ResponseEntity<>(categoryTransactions, HttpStatus.OK);
     }
 
     @GetMapping("/{id}")
-    public CategoryTransaction getCategoryTransaction(@PathVariable("id") int id) {
+    public ResponseEntity<CategoryTransaction> getCategoryTransaction(@PathVariable("id") int id) {
         Optional<CategoryTransaction> categoryTransaction = categoryTransactionsService.findById(id);
-        if (categoryTransaction.isPresent())
-            return categoryTransaction.get();
-        else
+        if (categoryTransaction.isEmpty())
             throw new NotFoundException("CategoryTransaction with this id wasn't found!");
+        return new ResponseEntity<>(categoryTransaction.get(), HttpStatus.OK);
     }
 
     @PostMapping("/add")
     public ResponseEntity<CategoryTransaction> addCategoryTransaction(@RequestBody @Valid CategoryTransaction categoryTransaction, BindingResult bindingResult) {
 
+        categoryTransaction.setId(0);
         if (bindingResult.hasErrors())
             returnDataErrorsToClient(bindingResult);
+
+        if (categoryTransaction.getPerson() == null)
+            throw new DataException("Person must not be null!");
+
+        Optional<Person> person = peopleService.findById(categoryTransaction.getPerson().getId());
+        if (person.isEmpty())
+            throw new NotFoundException("Person with this id wasn't found!");
+
+        if (categoryTransaction.getCategory() == null)
+            throw new DataException("Category must not be null!");
+
+        Optional<Category> category = categoriesService.findById(categoryTransaction.getCategory().getId());
+        if (category.isEmpty())
+            throw new NotFoundException("Category with this id wasn't found!");
+
+        if (categoryTransaction.getCreated_at() == null)
+            throw new DataException("Created_at must not be null!");
 
         categoryTransactionsService.save(categoryTransaction);
 
         return new ResponseEntity<>(categoryTransaction, HttpStatus.OK);
     }
 
-    @PostMapping("/update")
-    public CategoryTransaction updateCategoryTransaction(@RequestBody CategoryTransaction categoryTransaction) {
+    @PostMapping("/update/{id}")
+    public ResponseEntity<CategoryTransaction> updateCategoryTransaction(@PathVariable("id") int id, @RequestBody @Valid CategoryTransaction categoryTransaction, BindingResult bindingResult) {
+
+        categoryTransaction.setId(id);
+        if (bindingResult.hasErrors())
+            returnDataErrorsToClient(bindingResult);
+
+        Optional<CategoryTransaction> foundCategoryTransaction = categoryTransactionsService.findById(id);
+        if (foundCategoryTransaction.isEmpty())
+            throw new NotFoundException("CategoryTransaction with this id wasn't found!");
+
+        if (categoryTransaction.getPerson() == null)
+            throw new DataException("Person must not be null!");
+
+        Optional<Person> person = peopleService.findById(categoryTransaction.getPerson().getId());
+        if (person.isEmpty())
+            throw new NotFoundException("Person with this id wasn't found!");
+
+        if (categoryTransaction.getCategory() == null)
+            throw new DataException("Category must not be null!");
+
+        Optional<Category> category = categoriesService.findById(categoryTransaction.getCategory().getId());
+        if (category.isEmpty())
+            throw new NotFoundException("Category with this id wasn't found!");
+
+        if (categoryTransaction.getCreated_at() == null)
+            throw new DataException("Created_at must not be null!");
+
         categoryTransactionsService.update(categoryTransaction);
-        return categoryTransactionsService.findById(categoryTransaction.getId()).get();
+
+        return new ResponseEntity<>(categoryTransaction, HttpStatus.OK);
     }
 
+
     @PostMapping("/delete/{id}")
-    public void deleteCategoryTransaction(@PathVariable("id") int id) {
+    public ResponseEntity<CategoryTransaction> deleteCategoryTransaction(@PathVariable("id") int id) {
+
+        Optional<CategoryTransaction> foundCategoryTransaction = categoryTransactionsService.findById(id);
+        if (foundCategoryTransaction.isEmpty())
+            throw new NotFoundException("CategoryTransaction with this id wasn't found!");
+
         categoryTransactionsService.delete(id);
+
+        return new ResponseEntity<>(foundCategoryTransaction.get(), HttpStatus.OK);
     }
